Removes internal  folder usage: - Corrects imports - Moves files to root
