+++ api.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 	"os"
 
-	model "github.com/joelschutz/mercadopago-sdk-go/internal/models"
-	"github.com/joelschutz/mercadopago-sdk-go/internal/request"
+	model "github.com/joelschutz/mercadopago-sdk-go/models"
+	"github.com/joelschutz/mercadopago-sdk-go/request"
 )
 
 const BASEURL = "https://api.mercadopago.com"
